Move core test util tests to the right module
