+++ testing/test-utils/build.gradle
@@ -16,10 +16,6 @@ dependencies {
     }
     compile project(':client:mock')
 
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    compile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    compile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-
     compile "com.google.guava:guava:$guava_version"
 
     // Guava: Google test library (collections test suite)
